+++ typings/__init__.py
@@ -1,3 +1,3 @@
-from .constraints import VariableManager, ConstraintManager
+from .constraints import VariableManager, BaseConstraint, AuthorizationConstraint, SeparationOfDutyConstraint, BindingOfDutyConstraint, AtMostKConstraint, OneTeamConstraint, ConstraintManager
 from .solution import Solution, Verifier
 from .instance import Instance